nonebot_plugin_BAdrawcard/pyproject.toml
[tool.poetry]
name = "nonebot-plugin-badrawcard"
version = "0.1.0"
description = "模拟BA抽卡"
authors = ["LMZZ"]
readme = "README.md"
packages = [{include = "nonebot_plugin_badrawcard"}]

[tool.poetry.dependencies]
python = "^3.10"
redis = "^5.0.0"
pillow = "^10.0.1"
pydantic = {version = "^1.10.0", extras = ["dotenv"]}
aiofiles = "^23.2.1"
httpx = "^0.25.0"
loguru = "^0.7.2"
nonebot2 = "^2.1.3"
nonebot-adapter-onebot = "^2.3.1"


[tool.nonebot]
plugins = ["nonebot_plugin_apscheduler", "nonebot_plugin_saa"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
ignore = ["E402"]